requirements.txt
fastapi==0.104.0
Pillow==10.1.0
python-multipart==0.0.6
uvicorn==0.23.2 
